+++ RELEASE_NOTES.md
@@ -23,7 +23,8 @@ By deprecating these functions, we can simplify the library in 1.0, and focus mo
 * `RichTextCommand.ActionButtonGroup` has new inits.
 * `RichTextCommand.FormatMenu` is a lot more configurable.
 * `RichTextFormatToolbar` is now available on all platforms.
-* `RichTextFormatToolbar` has new configuration and style type.
+* `RichTextFormatToolbar` has new configuration and style types.
+* `RichTextFormatSidebar` has new configuration and style types.
 * `RichTextKeyboardToolbar` has a new config to always be shown.
 * `RichTextView` has a new theme that lets you define its style.
 * `RichTextViewComponent` has a new `hasRichTextStyle` function.